+++ docs/changelog.rst
@@ -7,6 +7,8 @@ v0.8.0 (in development)
 -----------------------
 - Drop support for Python 2.7, 3.4, and 3.5
 - Support Python 3.9
+- ``ensure_ascii`` parameter added to `PropertiesFile.dump()` and
+  `PropertiesFile.dumps()`
 
 v0.7.0 (2020-03-09)
 -------------------